reorg some screenshot stuff
